@@ -8,7 +8,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 
-import 'package:umeng_plugin_example/main.dart';
+import '../lib/main.dart';
+
 
 void main() {
   testWidgets('Verify Platform version', (WidgetTester tester) async {
@@ -19,7 +20,7 @@ void main() {
     expect(
       find.byWidgetPredicate(
         (Widget widget) => widget is Text &&
-                           widget.data.startsWith('Running on:'),
+                           widget.data!.startsWith('Running on:'),
       ),
       findsOneWidget,
     );
